Split out name and desc
